Fix issues accessing the ZHA gateway
The gateway could now also be a gateway proxy object that stores
the HASS object under `hass` instead of `_hass` and the HASS object
is no longer available in the non-proxy object since it lives in
the new ZHA (non-HA) library.

Having access to the proxy object is useful to get the `hass`
object, but accessing the application controller needs to be done
through the actual gateway. Therefore, a helper method has been
added with simple duck-type checking to reach the gateway through
the `gateway` property of the proxy.

Fixes: #251